refactor(connector-tester): move ChaosEngine into Engine/Chaos folder
